+++ x/crosschain/keeper/msg_migrate_tss_funds_test.go
@@ -23,7 +23,7 @@ func TestMsgServer_MigrateTssFunds(t *testing.T) {
 		msgServer := keeper.NewMsgServerImpl(*k)
 		chain := getValidEthChain(t)
 		amount := sdkmath.NewUint(100)
-		indexString, _ := setupTssMigrationParams(zk, k, ctx, *chain, amount, true)
+		indexString, _ := setupTssMigrationParams(zk, k, ctx, *chain, amount, true, true)
 		_, err := msgServer.MigrateTssFunds(ctx, &crosschaintypes.MsgMigrateTssFunds{
 			Creator: admin,
 			ChainId: chain.ChainId,
@@ -42,26 +42,27 @@ func TestMsgServer_MigrateTssFunds(t *testing.T) {
 		msgServer := keeper.NewMsgServerImpl(*k)
 		chain := getValidEthChain(t)
 		amount := sdkmath.NewUint(100)
-		indexString, _ := setupTssMigrationParams(zk, k, ctx, *chain, amount, false)
+		indexString, _ := setupTssMigrationParams(zk, k, ctx, *chain, amount, false, true)
 		_, err := msgServer.MigrateTssFunds(ctx, &crosschaintypes.MsgMigrateTssFunds{
 			Creator: admin,
 			ChainId: chain.ChainId,
 			Amount:  amount,
 		})
-		assert.ErrorIs(t, err, crosschaintypes.ErrUnableToUpdateTss)
+		assert.ErrorIs(t, err, crosschaintypes.ErrCannotMigrateTssFunds)
+		assert.ErrorContains(t, err, "no new tss address has been generated")
 		hash := crypto.Keccak256Hash([]byte(indexString))
 		index := hash.Hex()
 		_, found := k.GetCrossChainTx(ctx, index)
 		assert.False(t, found)
 	})
-	t.Run("unable to migrate funds pending cctx is present", func(t *testing.T) {
+	t.Run("unable to migrate funds when nonce low does not match nonce high", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
 		setAdminPolicies(ctx, zk, admin)
 		msgServer := keeper.NewMsgServerImpl(*k)
 		chain := getValidEthChain(t)
 		amount := sdkmath.NewUint(100)
-		indexString, tssPubkey := setupTssMigrationParams(zk, k, ctx, *chain, amount, false)
+		indexString, tssPubkey := setupTssMigrationParams(zk, k, ctx, *chain, amount, true, true)
 		k.SetPendingNonces(ctx, crosschaintypes.PendingNonces{
 			NonceLow:  1,
 			NonceHigh: 10,
@@ -70,17 +71,88 @@ func TestMsgServer_MigrateTssFunds(t *testing.T) {
 		})
 		_, err := msgServer.MigrateTssFunds(ctx, &crosschaintypes.MsgMigrateTssFunds{
 			Creator: admin,
-			ChainId: 1,
+			ChainId: chain.ChainId,
 			Amount:  amount,
 		})
-		assert.ErrorIs(t, err, crosschaintypes.ErrUnableToUpdateTss)
+		assert.ErrorIs(t, err, crosschaintypes.ErrCannotMigrateTssFunds)
+		assert.ErrorContains(t, err, "cannot migrate funds when there are pending nonces")
 		hash := crypto.Keccak256Hash([]byte(indexString))
 		index := hash.Hex()
 		_, found := k.GetCrossChainTx(ctx, index)
 		assert.False(t, found)
 	})
+	t.Run("unable to migrate funds when a pending cctx is presnt in migration info", func(t *testing.T) {
+		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
+		admin := sample.AccAddress()
+		setAdminPolicies(ctx, zk, admin)
+		msgServer := keeper.NewMsgServerImpl(*k)
+		chain := getValidEthChain(t)
+		amount := sdkmath.NewUint(100)
+		indexString, tssPubkey := setupTssMigrationParams(zk, k, ctx, *chain, amount, true, true)
+		k.SetPendingNonces(ctx, crosschaintypes.PendingNonces{
+			NonceLow:  1,
+			NonceHigh: 1,
+			ChainId:   chain.ChainId,
+			Tss:       tssPubkey,
+		})
+		existingCctx := sample.CrossChainTx(t, "sample_index")
+		existingCctx.CctxStatus.Status = crosschaintypes.CctxStatus_PendingOutbound
+		k.SetCrossChainTx(ctx, *existingCctx)
+		k.GetObserverKeeper().SetFundMigrator(ctx, observerTypes.TssFundMigratorInfo{
+			ChainId:            chain.ChainId,
+			MigrationCctxIndex: existingCctx.Index,
+		})
+		_, err := msgServer.MigrateTssFunds(ctx, &crosschaintypes.MsgMigrateTssFunds{
+			Creator: admin,
+			ChainId: chain.ChainId,
+			Amount:  amount,
+		})
+		assert.ErrorIs(t, err, crosschaintypes.ErrCannotMigrateTssFunds)
+		assert.ErrorContains(t, err, "cannot migrate funds while there are pending migrations")
+		hash := crypto.Keccak256Hash([]byte(indexString))
+		index := hash.Hex()
+		_, found := k.GetCrossChainTx(ctx, index)
+		assert.False(t, found)
+		_, found = k.GetCrossChainTx(ctx, existingCctx.Index)
+		assert.True(t, found)
+	})
+
+	t.Run("unable to migrate funds if current TSS is not present in TSSHistory and no new TSS has been generated", func(t *testing.T) {
+		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
+		admin := sample.AccAddress()
+		setAdminPolicies(ctx, zk, admin)
+		msgServer := keeper.NewMsgServerImpl(*k)
+		chain := getValidEthChain(t)
+		amount := sdkmath.NewUint(100)
+		indexString, _ := setupTssMigrationParams(zk, k, ctx, *chain, amount, false, false)
+		currentTss, found := k.GetObserverKeeper().GetTSS(ctx)
+		assert.True(t, found)
+		newTss := sample.Tss()
+		newTss.FinalizedZetaHeight = currentTss.FinalizedZetaHeight - 10
+		newTss.KeyGenZetaHeight = currentTss.KeyGenZetaHeight - 10
+		k.GetObserverKeeper().SetTSSHistory(ctx, newTss)
+		_, err := msgServer.MigrateTssFunds(ctx, &crosschaintypes.MsgMigrateTssFunds{
+			Creator: admin,
+			ChainId: chain.ChainId,
+			Amount:  amount,
+		})
+		assert.ErrorIs(t, err, crosschaintypes.ErrCannotMigrateTssFunds)
+		assert.ErrorContains(t, err, "current tss is the latest")
+		hash := crypto.Keccak256Hash([]byte(indexString))
+		index := hash.Hex()
+		_, found = k.GetCrossChainTx(ctx, index)
+		assert.False(t, found)
+	})
 }
-func setupTssMigrationParams(zk keepertest.ZetaKeepers, k *keeper.Keeper, ctx sdk.Context, chain common.Chain, amount sdkmath.Uint, setNewTss bool) (string, string) {
+func setupTssMigrationParams(
+	zk keepertest.ZetaKeepers,
+	k *keeper.Keeper,
+	ctx sdk.Context,
+	chain common.Chain,
+	amount sdkmath.Uint,
+	setNewTss bool,
+	setCurrentTSS bool,
+) (string, string) {
 	zk.ObserverKeeper.SetCrosschainFlags(ctx, observerTypes.CrosschainFlags{
 		IsInboundEnabled:  false,
 		IsOutboundEnabled: true,
@@ -98,7 +170,9 @@ func setupTssMigrationParams(zk keepertest.ZetaKeepers, k *keeper.Keeper, ctx sd
 	newTss.FinalizedZetaHeight = currentTss.FinalizedZetaHeight + 1
 	newTss.KeyGenZetaHeight = currentTss.KeyGenZetaHeight + 1
 	k.GetObserverKeeper().SetTSS(ctx, currentTss)
-	k.GetObserverKeeper().SetTSSHistory(ctx, currentTss)
+	if setCurrentTSS {
+		k.GetObserverKeeper().SetTSSHistory(ctx, currentTss)
+	}
 	if setNewTss {
 		k.GetObserverKeeper().SetTSSHistory(ctx, newTss)
 	}