+++ x/observer/types/keys.go
@@ -49,9 +49,10 @@ const (
 	BlockHeaderKey            = "BlockHeader-value-"
 	BlockHeaderStateKey       = "BlockHeaderState-value-"
 
-	BallotListKey = "BallotList-value-"
-	TSSKey        = "TSS-value-"
-	TSSHistoryKey = "TSS-History-value-"
+	BallotListKey      = "BallotList-value-"
+	TSSKey             = "TSS-value-"
+	TSSHistoryKey      = "TSS-History-value-"
+	TssFundMigratorKey = "FundsMigrator-value-"
 )
 
 func GetBlameIndex(chainID int64, nonce uint64, digest string, height uint64) string {